// src/types.ts
export type UserRole = "doctor" | "patient";